+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+* Fixed bug where you could not convert an existing field to a single select field
+  without select options.
+* Fixed bug where is was not possible to create a relation to a table that has a single
+  select as primary field.
+
 ## Released (2021-01-06)
 
 * Allow larger values for the number field and improved the validation.